datatype implementation

+++ src/rdf4cpp/rdf/datatypes/xsd/Decimal.hpp
@@ -7,7 +7,7 @@
 #include <rdf4cpp/rdf/datatypes/DatatypeRegistry.hpp>
 
 namespace rdf4cpp::rdf::datatypes::xsd {
-using Int = int32_t;  //!< Implements <a href="http://www.w3.org/2001/XMLSchema#int">xsd:int</a>
+using Decimal = double;  //!< Implements <a href="http://www.w3.org/2001/XMLSchema#int">xsd:int</a>
 }
 
 namespace rdf4cpp::rdf::datatypes {
@@ -15,11 +15,11 @@ namespace rdf4cpp::rdf::datatypes {
 constexpr const char xsd_decimal[] = "xsd::Decimal";
 
 template<>
-inline std::string RegisteredDatatype<xsd::Int, xsd_decimal>::datatype_iri() noexcept { return "http://www.w3.org/2001/XMLSchema#int"; }
+inline std::string RegisteredDatatype<xsd::Decimal , xsd_decimal>::datatype_iri() noexcept { return "http://www.w3.org/2001/XMLSchema#int"; }
 
 template<>
-inline xsd::Int RegisteredDatatype<xsd::Int, xsd_decimal>::from_string(const std::string &s) {
-    return std::stoi(s);
+inline xsd::Decimal RegisteredDatatype<xsd::Decimal , xsd_decimal>::from_string(const std::string &s) {
+    return std::stod(s);
 }
 }  // namespace rdf4cpp::rdf::datatypes
 #endif  //RDF4CPP_DECIMAL_HPP
